+++ include/aero/array/cpp_array.h
@@ -24,7 +24,7 @@ void aero_cpp_array_copy_in(aero_array_t *array, const aero_real_t *from);
 void aero_cpp_array_copy_out(aero_array_t *array, aero_real_t *to);
 
 /// Returns the number of elements in the array
-unsigned int aero_cpp_array_size(aero_array_t *array);
+size_t aero_cpp_array_size(aero_array_t *array);
 
 #ifdef __cplusplus
 } // extern "C"
